checking if the directory exists, if not, it will be created automatically
fix fun local uid and gid not match with prod bug
refactor related code
