+++ lib/client.js
@@ -1,4 +1,4 @@
-'use strict'
+'use strict';
 
 const osLocale = require('os-locale');
 const MNSClient = require('@alicloud/mns');
@@ -20,7 +20,7 @@ const getFcClient = async () => {
 
   FC.prototype.getAccountSettings = function(options = {}, headers = {}) {
     return this.get('/account-settings', options, headers);
-  }
+  };
 
   const fc = new FC(profile.accountId, {
     accessKeyID: profile.accessKeyId,
@@ -57,19 +57,18 @@ const getOtsPopClient = async () => {
 
 const getVpcPopClient = async () => {
   return await getPopClient('https://vpc.aliyuncs.com', '2016-04-28');
-}
+};
 
 const getEcsPopClient = async () => {
   return await getPopClient('https://ecs.aliyuncs.com', '2014-05-26');
-}
+};
 
 const getNasPopClient = async () => {
 
-  console.log("get nas pop client invoked");
   const profile = await getProfile();
 
   return await getPopClient(`http://nas.${profile.defaultRegion}.aliyuncs.com`, '2017-06-26');
-}
+};
 
 const getOtsClient = async (instanceName) => {
   const profile = await getProfile();
@@ -132,4 +131,4 @@ module.exports = {
   getVpcPopClient,
   getEcsPopClient,
   getNasPopClient
-}
+};
