+++ lib/deploy/deploy-by-tpl.js
@@ -4,9 +4,7 @@ const fs = require('fs');
 const path = require('path');
 const util = require('util');
 const yaml = require('js-yaml');
-const getProfile = require('../profile').getProfile;
-
-const { mark } = require('../profile');
+const { getProfile, mark } = require('../profile');
 
 const getTpl = require('../tpl').getTpl;
 const validate = require('../validate/validate');
@@ -17,21 +15,21 @@ const { green, yellow, red} = require('colors');
 const readFile = util.promisify(fs.readFile);
 const _ = require('lodash');
 
+const { makeService, makeFunction } = require('../fc');
+
+const { getTriggerNameList, makeTrigger } = require('../trigger');
+
 let {
   makeApi,
   makeApiTrigger,
-  makeFunction,
   makeGroup,
   makeOtsTable,
   makeOtsInstance,
-  makeService,
-  makeTrigger,
   makeMnsTopic,
   makeSlsProject,
   makeLogstore,
   makeLogstoreIndex,
-  makeCustomDomain,
-  getTriggerNameList
+  makeCustomDomain
 } = require('./deploy-support');
 
 let {
@@ -200,7 +198,7 @@ async function deployService(baseDir, serviceName, serviceRes) {
   // However, in some cases, users do not want to configure ram permissions for ram users.
   // https://github.com/aliyun/fun/issues/182
   // https://github.com/aliyun/fun/pull/223
-  if (!roleArn && (policies || !_.isEmpty(vpcConfig) || !_.isEmpty(logConfig))) {
+  if (!roleArn && (policies || !_.isEmpty(vpcConfig) || !_.isEmpty(logConfig) || !_.isEmpty(nasConfig))) {
     // create role
     role = await makeRole(roleName, createRoleIfNotExist);
   }
@@ -209,7 +207,7 @@ async function deployService(baseDir, serviceName, serviceRes) {
     await deployPolicies(serviceName, roleName, policies);
   }
 
-  if (!roleArn && !_.isEmpty(vpcConfig)) {
+  if (!roleArn && (!_.isEmpty(vpcConfig) || !_.isEmpty(nasConfig))) {
     await attachPolicyToRole('AliyunECSNetworkInterfaceManagementAccess', roleName);
   }
