+++ lib/deploy/deploy-support.js
@@ -2,8 +2,8 @@
 
 const fs = require('fs');
 const util = require('util');
-const vpc = require('../vpc');
-const nas = require('../nas');
+
+const { getCloudApiClient, getSlsClient, getMnsClient } = require('../client');
 
 const {
   getOtsClient,
@@ -11,37 +11,16 @@ const {
   getFcClient } = require('../client');
 
 const getProfile = require('../profile').getProfile;
-const zip = require('../package/zip');
 const debug = require('debug')('fun:deploy');
-const path = require('path');
 const { processApiParameters } = require('./deploy-support-api');
 
 const promiseRetry = require('../retry');
 
-const funignore = require('../package/ignore');
-
 const ram = require('../ram');
 
 const readFile = util.promisify(fs.readFile);
-const readdir = util.promisify(fs.readdir);
-const lstat = util.promisify(fs.lstat);
 
-const { green, red, yellow } = require('colors');
-const { addEnv } = require('../install/env');
-
-const _ = require('lodash');
-
-const defaultVpcConfig = {
-  securityGroupId: '',
-  vSwitchIds: [],
-  vpcId: ''
-};
-
-const defaultNasConfig = {
-  UserId: -1,
-  GroupId: -1,
-  MountPoints: []
-};
+const { green, red } = require('colors');
 
 async function makeLogstoreIndex(projectName, logstoreName) {
   const sls = await getSlsClient();
@@ -200,687 +179,6 @@ async function makeMnsTopic(topicName, properties) {
       retry(ex);
     }
   });
-}
-
-async function makeService({
-  serviceName,
-  role,
-  description,
-  internetAccess = true,
-  logConfig,
-  vpcConfig,
-  nasConfig
-}) {
-  const fc = await getFcClient();
-
-  var service;
-  await promiseRetry(async (retry, times) => {
-    try {
-      service = await fc.getService(serviceName);
-    } catch (ex) {
-
-      if (ex.code === 'AccessDenied') {
-
-        console.log(red(`\n${ex.message}`));
-
-        if (ex.message.indexOf('the caller is not authorized to perform') !== -1) {
-
-          console.error(red(`\nMaybe you need grant AliyunRAMFullAccess policy to the subuser or use the primary account. You can refer to Chinese doc https://github.com/aliyun/fun/blob/master/docs/usage/faq-zh.md#nopermissionerror-you-are-not-authorized-to-do-this-action-resource-acsramxxxxxxxxxxrole-action-ramgetrole or English doc https://github.com/aliyun/fun/blob/master/docs/usage/faq.md#nopermissionerror-you-are-not-authorized-to-do-this-action-resource-acsramxxxxxxxxxxrole-action-ramgetrole for help.\n\nIf you don’t want use the AliyunRAMFullAccess policy or primary account, you can also specify the Role property for Service. You can refer to Chinease doc https://github.com/aliyun/fun/blob/master/docs/specs/2018-04-03-zh-cn.md#aliyunserverlessservice or Enaglish doc https://github.com/aliyun/fun/blob/master/docs/specs/2018-04-03.md#aliyunserverlessservice for help.\n`));
-
-        } else if (ex.message.indexOf('FC service is not enabled for current user') !== -1) {
-
-          console.error(red(`\nFC service is not enabled for current user. Please enable FC service before using fun.\nYou can enable FC service on this page https://www.aliyun.com/product/fc .\n`));
-
-        } else {
-          console.error(red(`\nThe accountId you entered is incorrect. You can only use the primary account id, whether or not you use a sub-account or a primary account ak. You can get primary account ID on this page https://account.console.aliyun.com/#/secure .\n`));
-        }
-
-        throw ex;
-      } else if (ex.code !== 'ServiceNotFound') {
-        debug('error when getService, serviceName is %s, error is: \n%O', serviceName, ex);
-
-        console.log(red(`\tretry ${times} times`));
-        retry(ex);
-      }
-    }
-  });
-
-  const options = {
-    description,
-    role,
-    logConfig: {
-      project: logConfig.Project || '',
-      logstore: logConfig.Logstore || ''
-    }
-  };
-
-  if (internetAccess !== null) {
-    // vpc feature is not supported in some region
-    Object.assign(options, {
-      internetAccess
-    });
-  }
-
-  if (vpcConfig || _.isEqual(nasConfig, 'Auto')) {
-
-    if (_.isEqual(vpcConfig, 'Auto') || (_.isEmpty(vpcConfig) && _.isEqual(nasConfig, 'Auto'))) {
-      vpcConfig = await vpc.createDefaultVpcIfNotExist();
-
-      console.log("generated vpcConfig: " + vpcConfig);
-    }
-  }
-
-  Object.assign(options, {
-    vpcConfig: vpcConfig || defaultVpcConfig
-  });
-
-  if (nasConfig) {
-    if (_.isEqual(nasConfig, 'Auto')) {
-
-      const vpcId = vpcConfig.vpcId;
-      const vswitchId = _.head(vpcConfig.vswitchIds);
-
-      const mountPointDomain = await nas.createDefaultNasIfNotExist(vpcId, vswitchId);
-
-      nasConfig = {
-        UserId: 10007,
-        GroupId: 10007,
-        MountPoints: [
-          {
-            ServerAddr: `${mountPointDomain}:/`,
-            MountDir: '/mnt/nas/auto'
-          }
-        ]
-      }
-    }
-  }
-
-  Object.assign(options, {
-    nasConfig: nasConfig || defaultNasConfig
-  });
-
-  await promiseRetry(async (retry, times) => {
-    try {
-      if (!service) {
-        debug('create service %s, options is %j', serviceName, options);
-        service = await fc.createService(serviceName, options);
-      } else {
-        debug('update service %s, options is %j', serviceName, options);
-        service = await fc.updateService(serviceName, options);
-      }
-    } catch (ex) {
-      debug('error when createService or updateService, serviceName is %s, options is %j, error is: \n%O', serviceName, options, ex);
-
-      console.log(red(`\tretry ${times} times`));
-      retry(ex);
-    }
-  });
-
-  return service;
-}
-
-function generateFunIngore(baseDir, codeUri) {
-  const absCodeUri = path.resolve(codeUri);
-  const absBaseDir = path.resolve(baseDir);
-
-  const relative = path.relative(absBaseDir, absCodeUri);
-
-  if (codeUri.startsWith('..') || relative.startsWith('..')) {
-    console.warn(red(`\t\twarning: funignore is not supported for your CodeUri: ${codeUri}`));
-    return null;
-  }
-
-  return funignore(baseDir);
-}
-
-async function getFunCodeAsBase64(baseDir, codeUri, runtime, functionName) {
-  if (codeUri) {
-    if (codeUri.endsWith('.zip') || codeUri.endsWith('.jar')) {
-      return { base64: Buffer.from(await readFile(codeUri)).toString('base64') };
-    }
-  } else {
-    codeUri = './';
-  }
-
-  const ignore = generateFunIngore(baseDir, codeUri);
-
-  await detectLibrary(codeUri, runtime, baseDir, '\t\t', functionName);
-
-  return await zip.pack(codeUri, ignore);
-}
-
-function extractOssCodeUri(ossUri) {
-  const prefixLength = 'oss://'.length;
-
-  const index = ossUri.indexOf('/', prefixLength);
-
-  return {
-    ossBucketName: ossUri.substring(prefixLength, index),
-    ossObjectName: ossUri.substring(index + 1)
-  };
-}
-
-async function makeFunction(baseDir, {
-  serviceName,
-  functionName,
-  description = '',
-  handler,
-  initializer = '',
-  timeout = 3,
-  initializationTimeout = 3,
-  memorySize = 128,
-  runtime = 'nodejs6',
-  codeUri,
-  environmentVariables = {}
-}) {
-  const fc = await getFcClient();
-
-  var fn;
-  try {
-    fn = await fc.getFunction(serviceName, functionName);
-  } catch (ex) {
-    if (ex.code !== 'FunctionNotFound') {
-      throw ex;
-    }
-  }
-
-  let code;
-  // oss://my-bucket/function.zip
-  if (codeUri && codeUri.startsWith('oss://')) {
-    code = extractOssCodeUri(codeUri);
-  } else {
-    console.log(`\t\tWaiting for packaging function ${functionName} code...`);
-    const { base64, count } = await getFunCodeAsBase64(baseDir, codeUri, runtime, functionName);
-    const log = count ? ', the number of files you have packaged is：' + yellow(`${count}`) : '';
-    console.log(green(`\t\tpackage function ${functionName} code done` + log));
-    code = {
-      zipFile: base64
-    };
-  }
-
-  const params = {
-    description,
-    handler,
-    initializer,
-    timeout,
-    initializationTimeout,
-    memorySize,
-    runtime,
-    code,
-    environmentVariables: addEnv(environmentVariables)
-  };
-
-  for (let i in params.environmentVariables) {
-    if (!isNaN(params.environmentVariables[i])) {
-      debug(`the value in environmentVariables:${params.environmentVariables[i]} cast String Done`);
-      params.environmentVariables[i] = params.environmentVariables[i] + '';
-    }
-  }
-  if (!fn) {
-    // create
-    params['functionName'] = functionName;
-    fn = await fc.createFunction(serviceName, params);
-  } else {
-    // update
-    fn = await fc.updateFunction(serviceName, functionName, params);
-  }
-
-  return fn;
-}
-
-const triggerTypeMapping = {
-  'Datahub': 'datahub',
-  'Timer': 'timer',
-  'HTTP': 'http',
-  'Log': 'log',
-  'OSS': 'oss',
-  'RDS': 'rds',
-  'MNSTopic': 'mns_topic',
-  'TableStore': 'tablestore',
-  'CDN': 'cdn_events'
-};
-
-async function getSourceArn(triggerType, triggerProperties) {
-  const profile = await getProfile();
-
-  if (triggerType === 'Log') {
-    return `acs:log:${profile.defaultRegion}:${profile.accountId}:project/${triggerProperties.LogConfig.Project}`;
-  } else if (triggerType === 'RDS') {
-    return `acs:rds:${profile.defaultRegion}:${profile.accountId}:dbinstance/${triggerProperties.InstanceId}`;
-  } else if (triggerType === 'MNSTopic') {
-    if (triggerProperties.Region !== undefined) {
-      return `acs:mns:${triggerProperties.Region}:${profile.accountId}:/topics/${triggerProperties.TopicName}`;
-    }
-    return `acs:mns:${profile.defaultRegion}:${profile.accountId}:/topics/${triggerProperties.TopicName}`;
-  } else if (triggerType === 'TableStore') {
-    return `acs:ots:${profile.defaultRegion}:${profile.accountId}:instance/${triggerProperties.InstanceName}/table/${triggerProperties.TableName}`;
-  } else if (triggerType === 'OSS') {
-    return `acs:oss:${profile.defaultRegion}:${profile.accountId}:${triggerProperties.BucketName || triggerProperties.bucketName}`;
-  } else if (triggerType === 'CDN') {
-    return `acs:cdn:*:${profile.accountId}`;
-  }
-  return;
-}
-
-function getTriggerConfig(triggerType, triggerProperties) {
-  if (triggerType === 'Timer') {
-    return {
-      payload: triggerProperties.Payload,
-      cronExpression: triggerProperties.CronExpression,
-      enable: triggerProperties.Enable
-    };
-  } else if (triggerType === 'HTTP') {
-    return {
-      authType: (triggerProperties.AuthType).toLowerCase(),
-      methods: triggerProperties.Methods
-    };
-  } else if (triggerType === 'Log') {
-    const logConfig = triggerProperties.LogConfig;
-    const jobConfig = triggerProperties.JobConfig;
-    const sourceConfig = triggerProperties.SourceConfig;
-
-    return {
-      sourceConfig: {
-        logstore: sourceConfig.Logstore
-      },
-      jobConfig: {
-        maxRetryTime: jobConfig.MaxRetryTime,
-        triggerInterval: jobConfig.TriggerInterval
-      },
-      logConfig: {
-        project: logConfig.Project,
-        logstore: logConfig.Logstore,
-        functionParameter: logConfig.FunctionParameter || {}
-      },
-      Enable: triggerProperties.Enable || true
-    };
-  } else if (triggerType === 'RDS') {
-    return {
-      subscriptionObjects: triggerProperties.SubscriptionObjects,
-      retry: triggerProperties.Retry,
-      concurrency: triggerProperties.Concurrency,
-      eventFormat: triggerProperties.EventFormat
-    };
-  } else if (triggerType === 'MNSTopic') {
-    var notifyContentFormat = 'STREAM';
-    if (triggerProperties.NotifyContentFormat !== undefined) {
-      notifyContentFormat = triggerProperties.NotifyContentFormat;
-    }
-    var notifyStrategy = 'BACKOFF_RETRY';
-    if (triggerProperties.NotifyStrategy !== undefined) {
-      notifyStrategy = triggerProperties.NotifyStrategy;
-    }
-    var triggerCfg = {
-      NotifyContentFormat: notifyContentFormat,
-      NotifyStrategy: notifyStrategy
-    };
-    if (triggerProperties.FilterTag !== undefined) {
-      triggerCfg.FilterTag = triggerProperties.FilterTag;
-    }
-    return triggerCfg;
-  } else if (triggerType === 'TableStore') {
-    return {};
-  } else if (triggerType === 'OSS') {
-    return {
-      events: triggerProperties.Events || triggerProperties.events,
-      filter: triggerProperties.Filter || triggerProperties.filter
-    };
-  } else if (triggerType === 'CDN') {
-    return {
-      eventName: triggerProperties.EventName,
-      eventVersion: triggerProperties.EventVersion,
-      notes: triggerProperties.Notes,
-      filter: _.mapKeys(triggerProperties.Filter, (value, key) => {
-        return _.lowerFirst(key);
-      })
-    };
-  }
-  console.error(`trigger type is ${triggerType} not supported.`);
-}
-
-async function makeInvocationRole(serviceName, functionName, triggerType) {
-  if (triggerType === 'Log') {
-
-    const invocationRoleName = ram.normalizeRoleOrPoliceName(`AliyunFcGeneratedInvocationRole-${serviceName}-${functionName}`);
-
-    const invocationRole = await ram.makeRole(invocationRoleName, true, 'Used for fc invocation', {
-      'Statement': [{
-        'Action': 'sts:AssumeRole',
-        'Effect': 'Allow',
-        'Principal': {
-          'Service': [
-            'log.aliyuncs.com'
-          ]
-        }
-      }],
-      'Version': '1'
-    });
-
-    const policyName = ram.normalizeRoleOrPoliceName(`AliyunFcGeneratedInvocationPolicy-${serviceName}-${functionName}`);
-
-    await ram.makePolicy(policyName, {
-      'Version': '1',
-      'Statement': [{
-        'Action': [
-          'fc:InvokeFunction'
-        ],
-        'Resource': `acs:fc:*:*:services/${serviceName}/functions/*`,
-        'Effect': 'Allow'
-      },
-      {
-        'Action': [
-          'log:Get*',
-          'log:List*',
-          'log:PostLogStoreLogs',
-          'log:CreateConsumerGroup',
-          'log:UpdateConsumerGroup',
-          'log:DeleteConsumerGroup',
-          'log:ListConsumerGroup',
-          'log:ConsumerGroupUpdateCheckPoint',
-          'log:ConsumerGroupHeartBeat',
-          'log:GetConsumerGroupCheckPoint'
-        ],
-        'Resource': '*',
-        'Effect': 'Allow'
-      }
-      ]
-    });
-
-    await ram.attachPolicyToRole(policyName, invocationRoleName, 'Custom');
-    return invocationRole.Role;
-
-  } else if (triggerType === 'RDS' || triggerType === 'MNSTopic') {
-
-    const invocationRoleName = ram.normalizeRoleOrPoliceName(`FunCreateRole-${serviceName}-${functionName}`);
-    var tMap = {
-      'RDS': 'rds',
-      'MNSTopic': 'mns'
-    };
-    var principalService = util.format('%s.aliyuncs.com', tMap[triggerType]);
-
-    const invocationRole = await ram.makeRole(invocationRoleName, true, 'Used for fc invocation', {
-      'Statement': [{
-        'Action': 'sts:AssumeRole',
-        'Effect': 'Allow',
-        'Principal': {
-          'Service': [
-            principalService
-          ]
-        }
-      }],
-      'Version': '1'
-    });
-
-    const policyName = ram.normalizeRoleOrPoliceName(`FunCreatePolicy-${serviceName}-${functionName}`);
-
-    await ram.makePolicy(policyName, {
-      'Version': '1',
-      'Statement': [{
-        'Action': [
-          'fc:InvokeFunction'
-        ],
-        'Resource': `acs:fc:*:*:services/${serviceName}/functions/*`,
-        'Effect': 'Allow'
-      }]
-    });
-
-    await ram.attachPolicyToRole(policyName, invocationRoleName, 'Custom');
-
-    return invocationRole.Role;
-
-  } else if (triggerType === 'TableStore') {
-    const invocationRoleName = ram.normalizeRoleOrPoliceName(`FunCreateRole-${serviceName}-${functionName}`);
-
-    const invocationRole = await ram.makeRole(invocationRoleName, true, 'Used for fc invocation', {
-      'Statement': [{
-        'Action': 'sts:AssumeRole',
-        'Effect': 'Allow',
-        'Principal': {
-          'RAM': [
-            'acs:ram::1604337383174619:root'
-          ]
-        }
-      }],
-      'Version': '1'
-    });
-
-    const invkPolicyName = ram.normalizeRoleOrPoliceName(`FunCreateInvkPolicy-${serviceName}-${functionName}`);
-
-    await ram.makePolicy(invkPolicyName, {
-      'Version': '1',
-      'Statement': [{
-        'Action': [
-          'fc:InvokeFunction'
-        ],
-        'Resource': '*',
-        'Effect': 'Allow'
-      }]
-    });
-
-    await ram.attachPolicyToRole(invkPolicyName, invocationRoleName, 'Custom');
-
-    const otsReadPolicyName = ram.normalizeRoleOrPoliceName(`FunCreateOtsReadPolicy-${serviceName}-${functionName}`);
-
-    await ram.makePolicy(otsReadPolicyName, {
-      'Version': '1',
-      'Statement': [{
-        'Action': [
-          'ots:BatchGet*',
-          'ots:Describe*',
-          'ots:Get*',
-          'ots:List*'
-        ],
-        'Resource': '*',
-        'Effect': 'Allow'
-      }]
-    });
-
-    await ram.attachPolicyToRole(otsReadPolicyName, invocationRoleName, 'Custom');
-
-    return invocationRole.Role;
-  } else if (triggerType === 'OSS') {
-
-    const invocationRoleName = ram.normalizeRoleOrPoliceName(`FunCreateRole-${serviceName}-${functionName}`);
-
-    const invocationRole = await ram.makeRole(invocationRoleName, true, 'Used for fc invocation', {
-      'Statement': [
-        {
-          'Action': 'sts:AssumeRole',
-          'Effect': 'Allow',
-          'Principal': {
-            'Service': [
-              'oss.aliyuncs.com'
-            ]
-          }
-        }
-      ],
-      'Version': '1'
-    });
-
-    const policyName = ram.normalizeRoleOrPoliceName(`FunCreateOSSPolicy-${serviceName}-${functionName}`);
-
-    await ram.makePolicy(policyName, {
-      'Version': '1',
-      'Statement': [{
-        'Action': [
-          'fc:InvokeFunction'
-        ],
-        'Resource': `acs:fc:*:*:services/${serviceName}/functions/*`,
-        'Effect': 'Allow'
-      }]
-    });
-
-    await ram.attachPolicyToRole(policyName, invocationRoleName, 'Custom');
-    return invocationRole.Role;
-
-  } else if (triggerType === 'CDN') {
-
-    const invocationRoleName = ram.normalizeRoleOrPoliceName(`FunCreateRole-${serviceName}-${functionName}`);
-
-    const invocationRole = await ram.makeRole(invocationRoleName, true, 'Used for fc invocation', {
-      'Statement': [
-        {
-          'Action': 'sts:AssumeRole',
-          'Effect': 'Allow',
-          'Principal': {
-            'Service': [
-              'cdn.aliyuncs.com'
-            ]
-          }
-        }
-      ],
-      'Version': '1'
-    });
-
-    const policyName = ram.normalizeRoleOrPoliceName(`FunCreateCDNPolicy-${serviceName}-${functionName}`);
-
-    await ram.makePolicy(policyName, {
-      'Version': '1',
-      'Statement': [{
-        'Action': [
-          'fc:InvokeFunction'
-        ],
-        'Resource': `acs:fc:*:*:services/${serviceName}/functions/*`,
-        'Effect': 'Allow'
-      }]
-    });
-
-    await ram.attachPolicyToRole(policyName, invocationRoleName, 'Custom');
-    return invocationRole.Role;
-  }
-  return;
-}
-
-const runtimeTypeMapping = {
-  'nodejs6': 'node_modules',
-  'nodejs8': 'node_modules',
-  'python2.7': ['.egg-info', '.dist-info', '.fun'],
-  'python3': ['.egg-info', '.dist-info', '.fun'],
-  'php7.2': ['extension', 'vendor']
-};
-
-async function detectLibraryFolders(dirName, libraryFolders, childDir, wrap, functionName) {
-  if (Array.isArray(libraryFolders)) {
-    for (const iterator of libraryFolders) {
-      for (const name of childDir) {
-        if (_.endsWith(name, iterator)) {
-          console.warn(red(`${wrap}Fun detected that the library directory '${name}' is not included in function '${functionName}' CodeUri.\n\t\tPlease make sure if it is the right configuration. if yes, ignore please.`));
-          return;
-        }
-      }
-    }
-  } else {
-    if (childDir.includes(libraryFolders)) {
-      console.warn(red(`${wrap}Fun detected that the library directory '${libraryFolders}' is not included in function '${functionName}' CodeUri.\n\t\tPlease make sure if it is the right configuration. if yes, ignore please.`));
-    } else {
-
-      const funDir = childDir.filter(p => p === '.fun');
-      if (Array.isArray(funDir) && funDir.length > 0) {
-
-        const childFun = await readdir(path.join(dirName, '.fun'));
-
-        if (childFun.includes('root')) {
-
-          console.warn(red(`${wrap}Fun detected that the library directory '.fun/root' is not included in function '${functionName}' CodeUri.\n\t\tPlease make sure if it is the right configuration. if yes, ignore please.`));
-
-        }
-      }
-    }
-  }
-}
-
-async function detectLibrary(codeUri, runtime, baseDir, wrap = '', functionName) {
-  const stats = await lstat(codeUri);
-  if (stats.isFile()) {
-    let libraryFolders = runtimeTypeMapping[runtime];
-    let absoluteCodePath;
-
-    if (path.isAbsolute(codeUri)) {
-      absoluteCodePath = codeUri;
-    } else {
-      absoluteCodePath = path.join(baseDir, codeUri);
-    }
-
-    const dirName = path.dirname(absoluteCodePath);
-    const childDir = await readdir(dirName);
-
-    await detectLibraryFolders(dirName, libraryFolders, childDir, wrap, functionName);
-  }
-}
-
-async function getTriggerNameList({
-  serviceName,
-  functionName
-}) {
-  const fc = await getFcClient();
-  var listTriggerResponse = await fc.listTriggers(serviceName, functionName);
-
-  var triggerNameArray = [];
-
-  if (listTriggerResponse && listTriggerResponse.data.triggers) {
-    triggerNameArray = listTriggerResponse.data.triggers.map(p => p.triggerName);
-  }
-  return triggerNameArray;
-}
-
-async function makeTrigger({
-  serviceName,
-  functionName,
-  triggerName,
-  triggerType,
-  triggerProperties
-}) {
-  const fc = await getFcClient();
-  var trigger;
-  try {
-    trigger = await fc.getTrigger(serviceName, functionName, triggerName);
-  } catch (ex) {
-    if (ex.code !== 'TriggerNotFound') {
-      throw ex;
-    }
-  }
-
-  const params = {
-    triggerType: triggerTypeMapping[triggerType],
-    triggerConfig: getTriggerConfig(triggerType, triggerProperties)
-  };
-
-  debug('serviceName is %s, functionName is %s, trigger params is %j', serviceName, functionName, params);
-
-  let invocationRoleArn = triggerProperties.InvocationRole;
-
-  if (!invocationRoleArn) {
-    const invocationRole = await makeInvocationRole(serviceName, functionName, triggerType);
-    invocationRoleArn = invocationRole.Arn;
-  }
-
-  Object.assign(params, {
-    'invocationRole': invocationRoleArn
-  });
-
-  const sourceArn = await getSourceArn(triggerType, triggerProperties);
-  if (sourceArn) {
-    Object.assign(params, {
-      'sourceArn': sourceArn
-    });
-  }
-
-  if (!trigger) {
-    params.triggerName = triggerName;
-    trigger = await fc.createTrigger(serviceName, functionName, params);
-  } else {
-    if (triggerType === 'TableStore' || triggerType === 'MNSTopic') {
-      // no triggerConfig, so no updateTrigger, first delete, then create
-      // await fc.deleteTrigger(serviceName, functionName, triggerName);
-      // params.triggerName = triggerName;
-      // trigger = await fc.createTrigger(serviceName, functionName, params);
-      console.log(red(`\t\tWarning: TableStore and MNSTopic Trigger cann't update`));
-      return;
-    }
-    trigger = await fc.updateTrigger(serviceName, functionName, triggerName, params);
-  }
-
-  return trigger;
 }
 
 async function makeCustomDomain({
@@ -1189,8 +487,8 @@ async function makeOtsInstance(instanceName, clusterType, description) {
         ClusterType: clusterType,
         Description: description
       }, {
-          method: 'POST'
-        });
+        method: 'POST'
+      });
     } catch (ex) {
       if (ex.code === 'AuthFailed'
         || ex.code === 'InvalidParameter'
@@ -1254,20 +552,12 @@ async function makeOtsTable({
 module.exports = {
   makeApi,
   makeApiTrigger,
-  makeFunction,
   makeGroup,
   makeOtsTable,
   makeOtsInstance,
-  makeService,
-  makeTrigger,
   makeSlsProject,
   makeLogstore,
   makeLogstoreIndex,
   makeCustomDomain,
-  makeMnsTopic,
-  makeInvocationRole,
-  getFunCodeAsBase64,
-  getTriggerNameList,
-  getTriggerConfig,
-  detectLibrary
+  makeMnsTopic
 };
