+++ lib/docker.js
@@ -404,8 +404,8 @@ async function run(opts, event, outputStream, errorStream) {
 }
 
 function resolveDockerUser(nasConfig) {
-  let uid = 0;
-  let gid = 0;
+  let uid = 10007;
+  let gid = 10007;
 
   if (nasConfig) {
     const userId = nasConfig.UserId;
