+++ lib/exception-handler.js
@@ -6,7 +6,7 @@ const unrefTimeout = require('./unref-timeout');
 const handler = function (err) {
   // If the verbose option is true, in addition to the message,
   // print the stack of the error.
-  if (debug.enabled('*') && err.stack) {
+  if (err.stack) {
     console.error(err.stack);
   } else {
     console.error(err.message);
