+++ lib/fc.js
@@ -1,0 +1,492 @@
+'use strict';
+
+const { getFcClient } = require('./client');
+const vpc = require('./vpc');
+const nas = require('./nas');
+
+const util = require('util');
+const fs = require('fs');
+const readFile = util.promisify(fs.readFile);
+const path = require('path');
+const debug = require('debug')('fun:fc');
+const zip = require('./package/zip');
+const { green, red, yellow } = require('colors');
+const { addEnv } = require('./install/env');
+const funignore = require('./package/ignore');
+const lstat = util.promisify(fs.lstat);
+const readdir = util.promisify(fs.readdir);
+const _ = require('lodash');
+
+const promiseRetry = require('./retry');
+
+const FUN_GENERATED_SERVICE = 'fun-generated-default-service';
+
+const defaultVpcConfig = {
+  securityGroupId: '',
+  vSwitchIds: [],
+  vpcId: ''
+};
+
+const defaultNasConfig = {
+  UserId: -1,
+  GroupId: -1,
+  MountPoints: []
+};
+
+function generateFunIngore(baseDir, codeUri) {
+  const absCodeUri = path.resolve(codeUri);
+  const absBaseDir = path.resolve(baseDir);
+
+  const relative = path.relative(absBaseDir, absCodeUri);
+
+  if (codeUri.startsWith('..') || relative.startsWith('..')) {
+    console.warn(red(`\t\twarning: funignore is not supported for your CodeUri: ${codeUri}`));
+    return null;
+  }
+
+  return funignore(baseDir);
+}
+
+const runtimeTypeMapping = {
+  'nodejs6': 'node_modules',
+  'nodejs8': 'node_modules',
+  'python2.7': ['.egg-info', '.dist-info', '.fun'],
+  'python3': ['.egg-info', '.dist-info', '.fun'],
+  'php7.2': ['extension', 'vendor']
+};
+
+async function detectLibraryFolders(dirName, libraryFolders, childDir, wrap, functionName) {
+  if (Array.isArray(libraryFolders)) {
+    for (const iterator of libraryFolders) {
+      for (const name of childDir) {
+        if (_.endsWith(name, iterator)) {
+          console.warn(red(`${wrap}Fun detected that the library directory '${name}' is not included in function '${functionName}' CodeUri.\n\t\tPlease make sure if it is the right configuration. if yes, ignore please.`));
+          return;
+        }
+      }
+    }
+  } else {
+    if (childDir.includes(libraryFolders)) {
+      console.warn(red(`${wrap}Fun detected that the library directory '${libraryFolders}' is not included in function '${functionName}' CodeUri.\n\t\tPlease make sure if it is the right configuration. if yes, ignore please.`));
+    } else {
+
+      const funDir = childDir.filter(p => p === '.fun');
+      if (Array.isArray(funDir) && funDir.length > 0) {
+
+        const childFun = await readdir(path.join(dirName, '.fun'));
+
+        if (childFun.includes('root')) {
+
+          console.warn(red(`${wrap}Fun detected that the library directory '.fun/root' is not included in function '${functionName}' CodeUri.\n\t\tPlease make sure if it is the right configuration. if yes, ignore please.`));
+
+        }
+      }
+    }
+  }
+}
+
+async function detectLibrary(codeUri, runtime, baseDir, wrap = '', functionName) {
+  const stats = await lstat(codeUri);
+  if (stats.isFile()) {
+    let libraryFolders = runtimeTypeMapping[runtime];
+    let absoluteCodePath;
+
+    if (path.isAbsolute(codeUri)) {
+      absoluteCodePath = codeUri;
+    } else {
+      absoluteCodePath = path.join(baseDir, codeUri);
+    }
+
+    const dirName = path.dirname(absoluteCodePath);
+    const childDir = await readdir(dirName);
+
+    await detectLibraryFolders(dirName, libraryFolders, childDir, wrap, functionName);
+  }
+}
+
+function extractOssCodeUri(ossUri) {
+  const prefixLength = 'oss://'.length;
+
+  const index = ossUri.indexOf('/', prefixLength);
+
+  return {
+    ossBucketName: ossUri.substring(prefixLength, index),
+    ossObjectName: ossUri.substring(index + 1)
+  };
+}
+
+async function getFunCodeAsBase64(baseDir, codeUri, runtime, functionName) {
+  if (codeUri) {
+    if (codeUri.endsWith('.zip') || codeUri.endsWith('.jar')) {
+      return { base64: Buffer.from(await readFile(codeUri)).toString('base64') };
+    }
+  } else {
+    codeUri = './';
+  }
+
+  const ignore = generateFunIngore(baseDir, codeUri);
+
+  await detectLibrary(codeUri, runtime, baseDir, '\t\t', functionName);
+
+  return await zip.pack(codeUri, ignore);
+}
+
+async function makeFunction(baseDir, {
+  serviceName,
+  functionName,
+  description = '',
+  handler,
+  initializer = '',
+  timeout = 3,
+  initializationTimeout = 3,
+  memorySize = 128,
+  runtime = 'nodejs6',
+  codeUri,
+  environmentVariables = {}
+}) {
+  const fc = await getFcClient();
+
+  var fn;
+  try {
+    fn = await fc.getFunction(serviceName, functionName);
+  } catch (ex) {
+    if (ex.code !== 'FunctionNotFound') {
+      throw ex;
+    }
+  }
+
+  let code;
+  // oss://my-bucket/function.zip
+  if (codeUri && codeUri.startsWith('oss://')) {
+    code = extractOssCodeUri(codeUri);
+  } else {
+    console.log(`\t\tWaiting for packaging function ${functionName} code...`);
+    const { base64, count } = await getFunCodeAsBase64(baseDir, codeUri, runtime, functionName);
+    const log = count ? ', the number of files you have packaged is：' + yellow(`${count}`) : '';
+    console.log(green(`\t\tpackage function ${functionName} code done` + log));
+    code = {
+      zipFile: base64
+    };
+  }
+
+  const params = {
+    description,
+    handler,
+    initializer,
+    timeout,
+    initializationTimeout,
+    memorySize,
+    runtime,
+    code,
+    environmentVariables: addEnv(environmentVariables)
+  };
+
+  for (let i in params.environmentVariables) {
+    if (!isNaN(params.environmentVariables[i])) {
+      debug(`the value in environmentVariables:${params.environmentVariables[i]} cast String Done`);
+      params.environmentVariables[i] = params.environmentVariables[i] + '';
+    }
+  }
+  if (!fn) {
+    // create
+    params['functionName'] = functionName;
+    fn = await fc.createFunction(serviceName, params);
+  } else {
+    // update
+    fn = await fc.updateFunction(serviceName, functionName, params);
+  }
+
+  return fn;
+}
+
+async function makeService({
+  serviceName,
+  role,
+  description,
+  internetAccess = true,
+  logConfig = {},
+  vpcConfig,
+  nasConfig
+}) {
+  const fc = await getFcClient();
+
+  var service;
+  await promiseRetry(async (retry, times) => {
+    try {
+      service = await fc.getService(serviceName);
+    } catch (ex) {
+
+      if (ex.code === 'AccessDenied') {
+
+        console.log(red(`\n${ex.message}`));
+
+        if (ex.message.indexOf('the caller is not authorized to perform') !== -1) {
+
+          console.error(red(`\nMaybe you need grant AliyunRAMFullAccess policy to the subuser or use the primary account. You can refer to Chinese doc https://github.com/aliyun/fun/blob/master/docs/usage/faq-zh.md#nopermissionerror-you-are-not-authorized-to-do-this-action-resource-acsramxxxxxxxxxxrole-action-ramgetrole or English doc https://github.com/aliyun/fun/blob/master/docs/usage/faq.md#nopermissionerror-you-are-not-authorized-to-do-this-action-resource-acsramxxxxxxxxxxrole-action-ramgetrole for help.\n\nIf you don’t want use the AliyunRAMFullAccess policy or primary account, you can also specify the Role property for Service. You can refer to Chinease doc https://github.com/aliyun/fun/blob/master/docs/specs/2018-04-03-zh-cn.md#aliyunserverlessservice or Enaglish doc https://github.com/aliyun/fun/blob/master/docs/specs/2018-04-03.md#aliyunserverlessservice for help.\n`));
+
+        } else if (ex.message.indexOf('FC service is not enabled for current user') !== -1) {
+
+          console.error(red(`\nFC service is not enabled for current user. Please enable FC service before using fun.\nYou can enable FC service on this page https://www.aliyun.com/product/fc .\n`));
+
+        } else {
+          console.error(red(`\nThe accountId you entered is incorrect. You can only use the primary account id, whether or not you use a sub-account or a primary account ak. You can get primary account ID on this page https://account.console.aliyun.com/#/secure .\n`));
+        }
+
+        throw ex;
+      } else if (ex.code !== 'ServiceNotFound') {
+        debug('error when getService, serviceName is %s, error is: \n%O', serviceName, ex);
+
+        console.log(red(`\tretry ${times} times`));
+        retry(ex);
+      }
+    }
+  });
+
+  const options = {
+    description,
+    role,
+    logConfig: {
+      project: logConfig.Project || '',
+      logstore: logConfig.Logstore || ''
+    }
+  };
+
+  if (internetAccess !== null) {
+    // vpc feature is not supported in some region
+    Object.assign(options, {
+      internetAccess
+    });
+  }
+
+  if (!_.isEmpty(vpcConfig) || _.isEqual(nasConfig, 'Auto')) {
+
+    if (_.isEqual(vpcConfig, 'Auto') || (_.isEmpty(vpcConfig) && _.isEqual(nasConfig, 'Auto'))) {
+      vpcConfig = await vpc.createDefaultVpcIfNotExist();
+
+      debug('generated vpcConfig: %j', vpcConfig);
+    }
+  }
+
+  Object.assign(options, {
+    vpcConfig: vpcConfig || defaultVpcConfig
+  });
+
+  if (_.isEqual(nasConfig, 'Auto')) {
+
+    const vpcId = vpcConfig.vpcId;
+    const vswitchId = _.head(vpcConfig.vswitchIds);
+
+    console.log("\tusing 'NasConfig: Auto', will auto generate related nas file system");
+    nasConfig = await nas.generateAutoNasConfig(serviceName, vpcId, vswitchId);
+    console.log(green("\tgenerated auto NasConfig done: ", JSON.stringify(nasConfig)));
+  }
+
+  Object.assign(options, {
+    nasConfig: nasConfig || defaultNasConfig
+  });
+
+  await promiseRetry(async (retry, times) => {
+    try {
+      if (!service) {
+        debug('create service %s, options is %j', serviceName, options);
+        service = await fc.createService(serviceName, options);
+      } else {
+        debug('update service %s, options is %j', serviceName, options);
+        service = await fc.updateService(serviceName, options);
+      }
+    } catch (ex) {
+      debug('error when createService or updateService, serviceName is %s, options is %j, error is: \n%O', serviceName, options, ex);
+
+      console.log(red(`\tretry ${times} times`));
+      retry(ex);
+    }
+  });
+
+  // make sure nas dir exist
+  if (serviceName !== FUN_GENERATED_SERVICE
+    && !_.isEmpty(nasConfig)
+    && !_.isEmpty(nasConfig.MountPoints)) {
+
+    await ensureNasDirExist({
+      role, vpcConfig, nasConfig
+    });
+  }
+
+  return service;
+}
+
+function mapMountPointDir(mountPoints, func) {
+  let resolvedMountPoints = _.map(mountPoints, (mountPoint) => {
+    const serverAddr = mountPoint.ServerAddr;
+
+    const index = _.lastIndexOf(serverAddr, ':');
+    if (index >= 0) {
+      const mountPointDomain = serverAddr.substring(0, index);
+      const remoteDir = serverAddr.substring(index + 1);
+      const mountDir = mountPoint.MountDir;
+
+      debug('remoteDir is: %s', remoteDir);
+
+      return func(mountPointDomain, remoteDir, mountDir);
+    }
+  });
+
+  resolvedMountPoints = _.compact(resolvedMountPoints);
+
+  return resolvedMountPoints;
+}
+
+async function ensureNasDirExist({
+  role,
+  vpcConfig,
+  nasConfig
+}) {
+  const mountPoints = nasConfig.MountPoints;
+  const modifiedNasConfig = _.cloneDeep(nasConfig);
+
+  modifiedNasConfig.MountPoints = mapMountPointDir(mountPoints, (mountPointDomain, remoteDir, mountDir) => {
+    if (remoteDir !== '/') {
+      return {
+        ServerAddr: `${mountPointDomain}:/`,
+        MountDir: `${mountDir}`
+      };
+    } return null;
+  });
+
+  const dirsNeedToCheck = mapMountPointDir(mountPoints, (mountPointDomain, remoteDir, mountDir) => {
+    if (remoteDir !== '/') {
+      return path.join(mountDir, remoteDir);
+    } return null;
+  });
+
+  debug('dirs need to check: %s', dirsNeedToCheck);
+
+  if (!_.isEmpty(dirsNeedToCheck)) {
+    console.log(`\tChecking if nas directories ${dirsNeedToCheck} exists, if not, it will be created automatically`);
+
+    const utilFunctionName = await makeFcUtilsFunctionNasDirChecker(role, vpcConfig, modifiedNasConfig);
+
+    await invokeFcUtilsFunction({
+      functionName: utilFunctionName,
+      event: JSON.stringify(dirsNeedToCheck)
+    });
+  
+    console.log(green('\tChecking nas directories done', JSON.stringify(dirsNeedToCheck)));
+  } 
+}
+
+async function makeFcUtilsService(role, vpcConfig, nasConfig) {
+  return await makeService({
+    serviceName: FUN_GENERATED_SERVICE,
+    role,
+    description: 'generated by fun',
+    vpcConfig,
+    nasConfig
+  });
+}
+
+async function makeFcUtilsFunction({
+  serviceName,
+  functionName,
+  codes,
+  description = '',
+  handler,
+  timeout = 30,
+  memorySize = 128,
+  runtime = 'nodejs8'
+}) {
+  const fc = await getFcClient();
+
+  var fn;
+  try {
+    fn = await fc.getFunction(serviceName, functionName);
+  } catch (ex) {
+    if (ex.code !== 'FunctionNotFound') {
+      throw ex;
+    }
+  }
+
+  const base64 = await zip.packFromJson(codes);
+
+  let code = {
+    zipFile: base64
+  };
+
+  const params = {
+    description,
+    handler,
+    initializer: '',
+    timeout,
+    memorySize,
+    runtime,
+    code
+  };
+
+  if (!fn) {
+    // create
+    params['functionName'] = functionName;
+    fn = await fc.createFunction(serviceName, params);
+  } else {
+    // update
+    fn = await fc.updateFunction(serviceName, functionName, params);
+  }
+
+  return fn;
+}
+
+async function invokeFcUtilsFunction({
+  functionName,
+  event
+}) {
+  const fc = await getFcClient();
+  const rs = await fc.invokeFunction(FUN_GENERATED_SERVICE, functionName, event, {
+    'X-Fc-Log-Type': 'Tail'
+  });
+
+  if (rs.data !== 'OK') {
+    const log = rs.headers['x-fc-log-result'];
+
+    if (log) {
+      const decodedLog = Buffer.from(log, 'base64');
+
+      throw new Error(`fc utils function ${functionName} invoke error, error message is: ${decodedLog}`);
+    }
+  }
+}
+
+async function getFcUtilsFunctionCode(filename) {
+  const functionCode = await readFile(path.join(__dirname, 'fc-utils', filename));
+
+  const codes = {
+    'index.js': functionCode
+  };
+
+  return codes;
+}
+
+async function makeFcUtilsFunctionNasDirChecker(role, vpcConfig, nasConfig) {
+  await makeFcUtilsService(role, vpcConfig, nasConfig);
+
+  const functionName = 'nas_dir_checker';
+
+  const codes = await getFcUtilsFunctionCode('fc-nas-dir-check.js');
+
+  await makeFcUtilsFunction({
+    serviceName: FUN_GENERATED_SERVICE,
+    functionName: 'nas_dir_checker',
+    codes,
+    description: 'used for fun to ensure nas remote dir exist',
+    handler: 'index.handler'
+  });
+
+  return functionName;
+}
+
+module.exports = {
+  invokeFcUtilsFunction,
+  makeFcUtilsFunctionNasDirChecker,
+  FUN_GENERATED_SERVICE,
+  makeService,
+  makeFunction,
+  getFunCodeAsBase64
+};
