+++ lib/package/zip.js
@@ -8,6 +8,7 @@ const util = require('util');
 const debug = require('debug')('fun:zip');
 const JSZip = require('jszip');
 const glob = require('glob');
+const _ = require('lodash');
 
 const exists = util.promisify(fs.exists);
 const readdir = util.promisify(fs.readdir);
@@ -118,6 +119,27 @@ async function zipFolder(zip, folder, folders, funignore) {
   return count;
 }
 
+async function generateAsync(zip) {
+  return await zip.generateAsync({
+    type: 'base64',
+    platform: 'UNIX',
+    compression: 'DEFLATE',
+    compressionOptions: {
+      level: 9
+    }
+  });
+}
+
+exports.packFromJson = async function (files) {
+  const zip = new JSZip();
+
+  _.forEach(files, (content, path) => {
+    zip.file(path, content);
+  });
+
+  return await generateAsync(zip);
+};
+
 exports.pack = async function (file, funignore) {
   if (!(await exists(file))) {
     throw new Error('zip file %s is not exist.', file);
@@ -145,14 +167,8 @@ exports.pack = async function (file, funignore) {
   } else {
     throw new Error('file %s must be a regular file or directory.', file);
   }
-  const base64 = await zip.generateAsync({
-    type: 'base64',
-    platform: 'UNIX',
-    compression: 'DEFLATE',
-    compressionOptions: {
-      level: 9
-    }
-  });
+  const base64 = await generateAsync(zip);
 
   return { base64, count };
 };
+
