+++ lib/security-group.js
@@ -1,4 +1,4 @@
-'use strict'
+'use strict';
 
 var requestOption = {
   method: 'POST'
@@ -6,18 +6,14 @@ var requestOption = {
 
 async function describeSecurityGroups(client, region, vpcId, securityGroupName) {
   var params = {
-    "RegionId": region,
-    "VpcId": vpcId,
-    "SecurityGroupName": securityGroupName
+    'RegionId': region,
+    'VpcId': vpcId,
+    'SecurityGroupName': securityGroupName
   };
 
   const describeRs = await client.request('DescribeSecurityGroups', params, requestOption);
 
-  console.log("describeRs: " + JSON.stringify(describeRs));
-
   const securityGroup = describeRs.SecurityGroups.SecurityGroup;
-
-  console.log("security group: " + securityGroup);
 
   return securityGroup;
 }
@@ -25,21 +21,21 @@ async function describeSecurityGroups(client, region, vpcId, securityGroupName) 
 async function authDefaultSecurityGroupRules(ecsClient, region, securityGroupId) {
 
   const sgRules = [
-    { protocol: "TCP", port: "80/80" },
-    { protocol: "TCP", port: "443/443" },
-    { protocol: "ICMP", port: "-1/-1" },
-    { protocol: "TCP", port: "22/22" }
+    { protocol: 'TCP', port: '80/80' },
+    { protocol: 'TCP', port: '443/443' },
+    { protocol: 'ICMP', port: '-1/-1' },
+    { protocol: 'TCP', port: '22/22' }
   ];
 
   async function authSecurityGroupRule(ecsClient, region, securityGroupId, protocol, port) {
     var params = {
-      "RegionId": region,
-      "SecurityGroupId": securityGroupId,
-      "IpProtocol": protocol,
-      "PortRange": port,
-      "Policy": "Accept",
-      "SourceCidrIp": "0.0.0.0/0",
-      "NicType": "intranet"
+      'RegionId': region,
+      'SecurityGroupId': securityGroupId,
+      'IpProtocol': protocol,
+      'PortRange': port,
+      'Policy': 'Accept',
+      'SourceCidrIp': '0.0.0.0/0',
+      'NicType': 'intranet'
     };
 
     const rs = await ecsClient.request('AuthorizeSecurityGroup', params, requestOption);
@@ -53,15 +49,15 @@ async function authDefaultSecurityGroupRules(ecsClient, region, securityGroupId)
 
 async function createSecurityGroup(ecsClient, region, vpcId, securityGroupName) {
   var params = {
-    "RegionId": region,
-    "SecurityGroupName": securityGroupName,
-    "Description": 'default security group created by fc fun',
-    "VpcId": vpcId,
-    "SecurityGroupType": 'normal'
+    'RegionId': region,
+    'SecurityGroupName': securityGroupName,
+    'Description': 'default security group created by fc fun',
+    'VpcId': vpcId,
+    'SecurityGroupType': 'normal'
   };
 
   const createRs = await ecsClient.request('CreateSecurityGroup', params, requestOption);
-  console.log("createRs: " + JSON.stringify(createRs));
+  console.log('createRs: ' + JSON.stringify(createRs));
 
   const securityGroupId = createRs.SecurityGroupId;
 
@@ -73,4 +69,4 @@ async function createSecurityGroup(ecsClient, region, vpcId, securityGroupName) 
 module.exports = {
   describeSecurityGroups,
   createSecurityGroup
-}
+};
