+++ lib/validate/schema/service-resource.js
@@ -22,9 +22,9 @@ const serviceResourceSchema = {
           oneOf: [
             { 'type': 'string' },
             { '$ref': '/Resources/Service/Role' },
-            { 
-              'type': 'array', 
-              'items': { 
+            {
+              'type': 'array',
+              'items': {
                 oneOf: [
                   { 'type': 'string' },
                   { '$ref': '/Resources/Service/Role' }
@@ -37,47 +37,58 @@ const serviceResourceSchema = {
           'type': 'boolean'
         },
         'VpcConfig': {
-          'type': 'object',
-          'properties': {
-            'VpcId': {'type': 'string'},
-            'VSwitchIds': {
-              'type': 'array', 
-              'items': {'type': 'string'}
-            },
-            'SecurityGroupId': {'type': 'string'}
-          },
-          'required': ['VpcId', 'VSwitchIds', 'SecurityGroupId'],
-          'additionalProperties': false
+          oneOf: [
+            { 'type': 'string' },
+            {
+              'type': 'object',
+              'properties': {
+                'VpcId': { 'type': 'string' },
+                'VSwitchIds': {
+                  'type': 'array',
+                  'items': { 'type': 'string' }
+                },
+                'SecurityGroupId': { 'type': 'string' }
+              },
+              'required': ['VpcId', 'VSwitchIds', 'SecurityGroupId'],
+              'additionalProperties': false
+            }
+          ]
         },
         'LogConfig': {
           'type': 'object',
           'properties': {
-            'Project': {'type': 'string'},
-            'Logstore': {'type': 'string'}
+            'Project': { 'type': 'string' },
+            'Logstore': { 'type': 'string' }
           },
           'required': ['Project', 'Logstore'],
           'additionalProperties': false
         },
         'NasConfig': {
-          'type': 'object',
-          'properties': {
-            'UserId': {'type': 'integer'},
-            'GroupId': {'type': 'integer'},
-            'MountPoints': {
-              'type': 'array',
-              'items': {
-                'type': 'object',
-                'properties': {
-                  'ServerAddr': {'type': 'string'},
-                  'MountDir': {'type': 'string'}
-                },
-                'required': ['ServerAddr', 'MountDir'],
-                'additionalProperties': false
-              }
+          oneOf: [
+            { 'type': 'string' },
+            {
+              'type': 'object',
+              'properties': {
+                'UserId': { 'type': 'integer' },
+                'GroupId': { 'type': 'integer' },
+                'MountPoints': {
+                  'type': 'array',
+                  'items': {
+                    'type': 'object',
+                    'properties': {
+                      'ServerAddr': { 'type': 'string' },
+                      'MountDir': { 'type': 'string' }
+                    },
+                    'required': ['ServerAddr', 'MountDir'],
+                    'additionalProperties': false
+                  }
+                }
+              },
+              'required': ['UserId', 'GroupId', 'MountPoints'],
+              'additionalProperties': false
             }
-          },
-          'required': ['UserId', 'GroupId', 'MountPoints'],
-          'additionalProperties': false
+          ]
+
         }
       },
       'additionalProperties': false
