+++ lib/vpc.js
@@ -1,4 +1,4 @@
-'use strict'
+'use strict';
 
 const getProfile = require('./profile').getProfile;
 const _ = require('lodash');
@@ -26,9 +26,9 @@ async function findVpc(vpcClient, region, vpcName) {
 
   do {
     var params = {
-      "RegionId": region,
-      "PageSize": pageSize,
-      "PageNumber": ++requestPageNumber
+      'RegionId': region,
+      'PageSize': pageSize,
+      'PageNumber': ++requestPageNumber
     };
 
     const rs = await vpcClient.request('DescribeVpcs', params, requestOption);
@@ -43,25 +43,25 @@ async function findVpc(vpcClient, region, vpcName) {
   
     debug('find default vpc: %s', vpc);
 
-  } while(!vpc && totalCount && pageNumber && pageNumber * pageSize < totalCount);
+  } while (!vpc && totalCount && pageNumber && pageNumber * pageSize < totalCount);
 
   return vpc;
 }
 
 async function createVpc(vpcClient, region, vpcName) {
   var createParams = {
-    "RegionId": region,
-    "CidrBlock": "10.0.0.0/8",
-    "EnableIpv6": false,
-    "VpcName": vpcName,
-    "Description": "default vpc created by fc fun"
+    'RegionId': region,
+    'CidrBlock': '10.0.0.0/8',
+    'EnableIpv6': false,
+    'VpcName': vpcName,
+    'Description': 'default vpc created by fc fun'
   };
 
   const createRs = await vpcClient.request('CreateVpc', createParams, requestOption);
 
   const vpcId = createRs.VpcId;
 
-  debug("create vpc rs is: %j", createRs);
+  debug('create vpc rs is: %j', createRs);
 
   await waitVpcUntilAvaliable(vpcClient, region, vpcId);
 
@@ -81,8 +81,8 @@ async function waitVpcUntilAvaliable(vpcClient, region, vpcId) {
     count++;
 
     var params = {
-      "RegionId": region,
-      "VpcId": vpcId
+      'RegionId': region,
+      'VpcId': vpcId
     };
     
     await sleep(800);
@@ -91,9 +91,9 @@ async function waitVpcUntilAvaliable(vpcClient, region, vpcId) {
     
     status = rs.Vpcs.Vpc[0].Status;
 
-    debug("vpc status is: " + status);
+    debug('vpc status is: ' + status);
     
-  } while(count < 10 && status !== 'Available');
+  } while (count < 10 && status !== 'Available');
 }
 
 async function createDefaultVSwitchIfNotExist(vpcClient, region, vpcId, vswitchIds) {
@@ -109,17 +109,14 @@ async function createDefaultVSwitchIfNotExist(vpcClient, region, vpcId, vswitchI
 async function createDefaultSecurityGroupIfNotExist(ecsClient, region, vpcId) {
   // check fun default security group exist?
   const defaultSecurityGroup = await securityGroup.describeSecurityGroups(ecsClient, region, vpcId, defaultSecurityGroupName);
-  console.log("default security grpup: " + JSON.stringify(defaultSecurityGroup));
-
-  // create security group if not exist
-  console.log('check security');
+  debug('default security grpup: %j', defaultSecurityGroup);
 
   // create security group
   if (_.isEmpty(defaultSecurityGroup)) {
     return await securityGroup.createSecurityGroup(ecsClient, region, vpcId, defaultSecurityGroupName);
-  } else {
-    return defaultSecurityGroup[0].SecurityGroupId;
-  }
+  } 
+  return defaultSecurityGroup[0].SecurityGroupId;
+  
 }
 
 async function createDefaultVpcIfNotExist() {
@@ -145,7 +142,7 @@ async function createDefaultVpcIfNotExist() {
     vpcId = await createVpc(vpcClient, region, defaultVpcName);
   }
 
-  console.log("vpcId is " + vpcId);
+  debug('vpcId is %s', vpcId);
 
   const vswitchId = await createDefaultVSwitchIfNotExist(vpcClient, region, vpcId, vswitchIds);
 
@@ -159,7 +156,7 @@ async function createDefaultVpcIfNotExist() {
     vpcId,
     vswitchIds,
     securityGroupId
-  }
+  };
 }
 
 module.exports = {
