+++ lib/vswitch.js
@@ -1,4 +1,4 @@
-'use strict'
+'use strict';
 
 const _ = require('lodash');
 
@@ -18,7 +18,7 @@ async function createDefaultVSwitch(vpcClient, region, vpcId, vswitchName) {
     region,
     vpcId,
     zoneId: vswitchZoneId,
-    vswitchName: vswitchName,
+    vswitchName: vswitchName
   });
 
   return vswitchId;
@@ -26,8 +26,8 @@ async function createDefaultVSwitch(vpcClient, region, vpcId, vswitchName) {
 
 async function getVSwitchName(vpcClient, region, vswitchId) {
   var params = {
-    "RegionId": region,
-    "VSwitchId": vswitchId
+    'RegionId': region,
+    'VSwitchId': vswitchId
   };
 
 
@@ -46,14 +46,14 @@ async function findVswitchExistByName(vpcClient, region, vswitchIds, searchVSwti
       const vswitchName = await getVSwitchName(vpcClient, region, vswitchId);
 
       if (_.isEqual(searchVSwtichName, vswitchName)) {
-        debug("found default vswitchId: " + vswitchId);
+        debug('found default vswitchId: ' + vswitchId);
 
         return vswitchId;
       }
     }
   }
 
-  debug("could not find %s from %j for region %s", searchVSwtichName, vswitchIds, region);
+  debug('could not find %s from %j for region %s', searchVSwtichName, vswitchIds, region);
 
   return null;
 }
@@ -65,15 +65,15 @@ async function createVSwitch(vpcClient, {
   vswitchName
 }) {
   var params = {
-    "RegionId": region,
-    "VpcId": vpcId,
-    "ZoneId": zoneId,
-    "CidrBlock": "10.20.0.0/16",
-    "VSwitchName": vswitchName,
-    "Description": 'default vswitch created by fc fun'
+    'RegionId': region,
+    'VpcId': vpcId,
+    'ZoneId': zoneId,
+    'CidrBlock': '10.20.0.0/16',
+    'VSwitchName': vswitchName,
+    'Description': 'default vswitch created by fc fun'
   };
 
-  debug("createVSwitch params is %j", params);
+  debug('createVSwitch params is %j', params);
 
   const createRs = await vpcClient.request('CreateVSwitch', params, requestOption);
 
@@ -96,7 +96,7 @@ async function getFcAllowedZones() {
 
   const fcAllowedZones = fcRs.data.availableAZs;
 
-  debug("fc allowed zones: %j", fcAllowedZones);
+  debug('fc allowed zones: %j', fcAllowedZones);
 
   return fcAllowedZones;
 }
@@ -109,17 +109,17 @@ async function selectAllowedVSwitchZone(vpcClient, region) {
   const usedZoneId = await selectVSwitchZoneId(fcAllowedZones, zones);
 
   if (!usedZoneId) {
-    throw new Error("no availiable zone for vswitch");
+    throw new Error('no availiable zone for vswitch');
   }
 
-  debug("select allowed switch zone: ", usedZoneId);
+  debug('select allowed switch zone: ', usedZoneId);
 
   return usedZoneId;
 }
 
 async function describeZones(vpcClient, region) {
   const params = {
-    "RegionId": region
+    'RegionId': region
   };
 
   const zones = await vpcClient.request('DescribeZones', params, requestOption);
