+++ test/deploy/deploy-support.test.js
@@ -4,18 +4,7 @@ var nock = require('nock');
 
 let deploySupport = require('../../lib/deploy/deploy-support');
 
-const ram = require('../../lib/ram');
 const { setProcess } = require('../test-utils');
-const proxyquire = require('proxyquire');
-const FC = require('@alicloud/fc2');
-const sinon = require('sinon');
-const sandbox = sinon.createSandbox();
-const assert = sinon.assert;
-const zip = require('../../lib/package/zip');
-const expect = require('expect.js');
-const path = require('path');
-const util = require('util');
-
 
 describe('make', () => {
 
@@ -309,465 +298,4 @@ describe('make', () => {
   });
 
 });
-describe('Incorrect environmental variables', ()=> {
-  let restoreProcess;
 
-  beforeEach(async () => {
-    sandbox.stub(FC.prototype, 'getFunction').resolves({});
-    sandbox.stub(FC.prototype, 'updateFunction').resolves({});
-    sandbox.stub(zip, 'pack').resolves({});
-
-    deploySupport = await proxyquire('../../lib/deploy/deploy-support', {
-      '../package/zip': zip,
-      '@alicloud/fc2': FC
-    });
-
-    restoreProcess = setProcess({
-      ACCOUNT_ID: 'ACCOUNT_ID',
-      ACCESS_KEY_ID: 'ACCESS_KEY_ID',
-      ACCESS_KEY_SECRET: 'ACCESS_KEY_SECRET',
-      DEFAULT_REGION: 'cn-shanghai'
-    });
-  });
-
-  afterEach(() => {
-    sandbox.restore();
-    restoreProcess();
-  });
-
-  it('should cast env value to String', async ()=> {
-    await deploySupport.makeFunction(path.join('examples', 'local'), {
-      serviceName: 'localdemo',
-      functionName: 'nodejs6',
-      description: 'Hello world with nodejs6!',
-      handler: 'index.handler',
-      initializer: null,
-      timeout: 3,
-      initializationTimeout: 3,
-      memorySize: 128,
-      runtime: 'nodejs6',
-      codeUri: path.join('examples', 'local', 'nodejs6'),
-      environmentVariables: {'StringTypeValue1': 123, 'StringTypeValue2': 'test'}
-    });    
-    
-    assert.calledWith(
-      FC.prototype.updateFunction,
-      'localdemo',
-      'nodejs6',
-      {
-        description: 'Hello world with nodejs6!',
-        handler: 'index.handler',
-        initializer: null,
-        timeout: 3,
-        initializationTimeout: 3,
-        memorySize: 128,
-        runtime: 'nodejs6',
-        code: {
-          zipFile: undefined
-        },
-        environmentVariables: {
-          StringTypeValue1: '123',
-          StringTypeValue2: 'test',
-          LD_LIBRARY_PATH: '/code/.fun/root/usr/lib:/code/.fun/root/usr/lib/x86_64-linux-gnu:/code:/code/lib:/usr/local/lib',
-          PATH: '/code/.fun/root/usr/local/bin:/code/.fun/root/usr/local/sbin:/code/.fun/root/usr/bin:/code/.fun/root/usr/sbin:/code/.fun/root/sbin:/code/.fun/root/bin:/code/.fun/python/bin:/usr/local/bin:/usr/local/sbin:/usr/bin:/usr/sbin:/sbin:/bin',
-          PYTHONUSERBASE: '/code/.fun/python'
-        }
-      });
-  });
-});
-
-describe('make invocation role', () => {
-
-  let restoreProcess;
-
-  beforeEach(async () => {
-
-    sandbox.stub(ram, 'makeRole').resolves({
-      Role: 'generated-role-name'
-    });
-    sandbox.stub(ram, 'makePolicy').resolves({});
-    sandbox.stub(ram, 'attachPolicyToRole').resolves({});
-
-    deploySupport = await proxyquire('../../lib/deploy/deploy-support', {
-      '../../lib/ram': ram
-    });
-
-    restoreProcess = setProcess({
-      ACCOUNT_ID: 'testAccountId',
-      ACCESS_KEY_ID: 'testKeyId',
-      ACCESS_KEY_SECRET: 'testKeySecret',
-      DEFAULT_REGION: 'cn-shanghai'
-    });
-  });
-
-  afterEach(() => {
-    sandbox.restore();
-    restoreProcess();
-  });
-
-  it('makeInvocationRole of log', async () => {
-    const role = await deploySupport.makeInvocationRole('test-service_name', 'test-function_name', 'Log');
-
-    assert.calledOnce(ram.makeRole);
-    assert.calledWith(ram.makeRole,
-      'AliyunFcGeneratedInvocationRole-test-service-name-test-function-name',
-      true,
-      'Used for fc invocation',
-      {
-        Statement: [{
-          Action: 'sts:AssumeRole',
-          Effect: 'Allow',
-          Principal: { Service: ['log.aliyuncs.com'] }
-        }],
-        Version: '1'
-      });
-
-    assert.calledOnce(ram.makePolicy);
-    assert.calledWith(ram.makePolicy, 'AliyunFcGeneratedInvocationPolicy-test-service-name-test-function-name',
-      {
-        Statement: [{
-          Action: ['fc:InvokeFunction'],
-          Effect: 'Allow',
-          Resource: 'acs:fc:*:*:services/test-service_name/functions/*'
-        }, {
-          Action: ['log:Get*', 'log:List*', 'log:PostLogStoreLogs', 'log:CreateConsumerGroup', 'log:UpdateConsumerGroup', 'log:DeleteConsumerGroup', 'log:ListConsumerGroup', 'log:ConsumerGroupUpdateCheckPoint', 'log:ConsumerGroupHeartBeat', 'log:GetConsumerGroupCheckPoint'],
-          Effect: 'Allow',
-          Resource: '*'
-        }],
-        Version: '1'
-      });
-
-    assert.calledOnce(ram.attachPolicyToRole);
-    assert.calledWith(ram.attachPolicyToRole, 'AliyunFcGeneratedInvocationPolicy-test-service-name-test-function-name', 'AliyunFcGeneratedInvocationRole-test-service-name-test-function-name', 'Custom');
-
-    expect(role).to.be('generated-role-name');
-  });
-
-  it('makeInvocationRole of rds', async () => {
-    const role = await deploySupport.makeInvocationRole('test-service_name', 'test-function_name', 'RDS');
-
-    assert.calledOnce(ram.makeRole);
-    assert.calledWith(ram.makeRole,
-      'FunCreateRole-test-service-name-test-function-name',
-      true,
-      'Used for fc invocation',
-      {
-        Statement: [{
-          Action: 'sts:AssumeRole',
-          Effect: 'Allow',
-          Principal: { Service: ['rds.aliyuncs.com'] }
-        }],
-        Version: '1'
-      });
-
-    assert.calledOnce(ram.makePolicy);
-    assert.calledWith(ram.makePolicy, 'FunCreatePolicy-test-service-name-test-function-name',
-      {
-        Statement: [{
-          Action: ['fc:InvokeFunction'],
-          Effect: 'Allow',
-          Resource: 'acs:fc:*:*:services/test-service_name/functions/*'
-        }],
-        Version: '1'
-      });
-
-    assert.calledOnce(ram.attachPolicyToRole);
-    assert.calledWith(ram.attachPolicyToRole, 'FunCreatePolicy-test-service-name-test-function-name', 'FunCreateRole-test-service-name-test-function-name', 'Custom');
-
-    expect(role).to.be('generated-role-name');
-  });
-
-  it('makeInvocationRole of mns topic', async () => {
-    const role = await deploySupport.makeInvocationRole('test-service_name', 'test-function_name', 'MNSTopic');
-
-    assert.calledOnce(ram.makeRole);
-    assert.calledWith(ram.makeRole,
-      'FunCreateRole-test-service-name-test-function-name',
-      true,
-      'Used for fc invocation',
-      {
-        Statement: [{
-          Action: 'sts:AssumeRole',
-          Effect: 'Allow',
-          Principal: { Service: ['mns.aliyuncs.com'] }
-        }],
-        Version: '1'
-      });
-
-    assert.calledOnce(ram.makePolicy);
-    assert.calledWith(ram.makePolicy, 'FunCreatePolicy-test-service-name-test-function-name',
-      {
-        Statement: [{
-          Action: ['fc:InvokeFunction'],
-          Effect: 'Allow',
-          Resource: 'acs:fc:*:*:services/test-service_name/functions/*'
-        }],
-        Version: '1'
-      });
-
-    assert.calledOnce(ram.attachPolicyToRole);
-    assert.calledWith(ram.attachPolicyToRole, 'FunCreatePolicy-test-service-name-test-function-name', 'FunCreateRole-test-service-name-test-function-name', 'Custom');
-
-    expect(role).to.be('generated-role-name');
-  });
-
-  it('makeInvocationRole of table store', async () => {
-    const role = await deploySupport.makeInvocationRole('test-service_name', 'test-function_name', 'TableStore');
-
-    assert.calledOnce(ram.makeRole);
-    assert.calledWith(ram.makeRole,
-      'FunCreateRole-test-service-name-test-function-name',
-      true,
-      'Used for fc invocation',
-      {
-        Statement: [{
-          Action: 'sts:AssumeRole',
-          Effect: 'Allow',
-          Principal: { RAM: ['acs:ram::1604337383174619:root'] } }],
-        Version: '1'
-      });
-
-    assert.calledTwice(ram.makePolicy);
-    assert.calledWith(ram.makePolicy.firstCall, 'FunCreateInvkPolicy-test-service-name-test-function-name',
-      {
-        Statement: [{ Action: ['fc:InvokeFunction'], Effect: 'Allow', Resource: '*' }],
-        Version: '1'
-      });
-    assert.calledWith(ram.makePolicy.secondCall, 'FunCreateOtsReadPolicy-test-service-name-test-function-name',
-      {
-        Statement: [{
-          Action: ['ots:BatchGet*', 'ots:Describe*', 'ots:Get*', 'ots:List*'],
-          Effect: 'Allow',
-          Resource: '*'
-        }],
-        Version: '1'
-      });
-
-    assert.calledTwice(ram.attachPolicyToRole);
-    assert.calledWith(ram.attachPolicyToRole.firstCall, 'FunCreateInvkPolicy-test-service-name-test-function-name', 'FunCreateRole-test-service-name-test-function-name', 'Custom');
-    assert.calledWith(ram.attachPolicyToRole.secondCall, 'FunCreateOtsReadPolicy-test-service-name-test-function-name', 'FunCreateRole-test-service-name-test-function-name', 'Custom');
-
-    expect(role).to.be('generated-role-name');
-  });
-
-  it('makeInvocationRole of oss', async () => {
-    const role = await deploySupport.makeInvocationRole('oss-service_name', 'oss-function_name', 'OSS');
-
-    assert.calledOnce(ram.makeRole);
-    assert.calledWith(ram.makeRole,
-      'FunCreateRole-oss-service-name-oss-function-name',
-      true,
-      'Used for fc invocation',
-      {
-        Statement: [
-          {
-            'Action': 'sts:AssumeRole',
-            'Effect': 'Allow',
-            'Principal': {
-              'Service': [
-                'oss.aliyuncs.com'
-              ]
-            }
-          }
-        ],
-        Version: '1'
-      });
-
-    assert.calledWith(ram.makePolicy, 'FunCreateOSSPolicy-oss-service-name-oss-function-name',
-      {
-        Statement: [{
-          'Action': [
-            'fc:InvokeFunction'
-          ],
-          'Resource': `acs:fc:*:*:services/oss-service_name/functions/*`,
-          'Effect': 'Allow'
-        }],
-        Version: '1'
-      });
-    assert.calledOnce(ram.attachPolicyToRole);
-    assert.calledWith(ram.attachPolicyToRole, 'FunCreateOSSPolicy-oss-service-name-oss-function-name', 'FunCreateRole-oss-service-name-oss-function-name', 'Custom');
-
-    expect(role).to.be('generated-role-name');
-  });
-
-  it('makeInvocationRole of cdn', async () => {
-    const role = await deploySupport.makeInvocationRole('cdn-service_name', 'cdn-function_name', 'CDN');
-
-    assert.calledOnce(ram.makeRole);
-    assert.calledWith(ram.makeRole,
-      'FunCreateRole-cdn-service-name-cdn-function-name',
-      true,
-      'Used for fc invocation',
-      {
-        Statement: [
-          {
-            'Action': 'sts:AssumeRole',
-            'Effect': 'Allow',
-            'Principal': {
-              'Service': [
-                'cdn.aliyuncs.com'
-              ]
-            }
-          }
-        ],
-        Version: '1'
-      });
-
-    assert.calledWith(ram.makePolicy, 'FunCreateCDNPolicy-cdn-service-name-cdn-function-name',
-      {
-        Statement: [{
-          'Action': [
-            'fc:InvokeFunction'
-          ],
-          'Resource': `acs:fc:*:*:services/cdn-service_name/functions/*`,
-          'Effect': 'Allow'
-        }],
-        Version: '1'
-      });
-    assert.calledOnce(ram.attachPolicyToRole);
-    assert.calledWith(ram.attachPolicyToRole, 'FunCreateCDNPolicy-cdn-service-name-cdn-function-name', 'FunCreateRole-cdn-service-name-cdn-function-name', 'Custom');
-
-    expect(role).to.be('generated-role-name');
-  });
-});
-
-describe('test getFunCodeAsBase64', () => {
-
-  let restoreProcess;
-  let lstat;
-
-  beforeEach(async () => {
-
-    sandbox.stub(zip, 'pack').resolves({});
-
-    lstat = sandbox.stub();
-
-    lstat.resolves({
-      isFile: function () { return false; }
-    });
-
-    sandbox.stub(util, 'promisify').returns(lstat);
-
-    deploySupport = await proxyquire('../../lib/deploy/deploy-support', {
-      '../package/zip': zip,
-      'util': util
-    });
-
-    restoreProcess = setProcess({
-      ACCOUNT_ID: 'testAccountId',
-      ACCESS_KEY_ID: 'testKeyId',
-      ACCESS_KEY_SECRET: 'testKeySecret'
-    });
-  });
-
-  afterEach(() => {
-    sandbox.restore();
-    restoreProcess();
-  });
-
-  it('test getFunCodeAsBase64: codeUri outside baseDir', async () => {
-    await deploySupport.getFunCodeAsBase64('/a/b', '/a');
-    assert.calledWith(lstat, '/a');
-    assert.calledWith(zip.pack, '/a', null);
-  });
-
-
-  it('test getFunCodeAsBase64: codeUri outside baseDir2', async () => {
-    await deploySupport.getFunCodeAsBase64('/a/b', '../');
-    assert.calledWith(lstat, '../');
-    assert.calledWith(zip.pack, '../', null);
-  });
-
-  it('test getFunCodeAsBase64: codeUri within baseDir', async () => {
-    await deploySupport.getFunCodeAsBase64('/a/b', '/a/b/c');
-    assert.calledWith(lstat, '/a/b/c');
-    assert.calledWith(zip.pack, '/a/b/c', sinon.match.func);
-  });
-
-
-  it('test getFunCodeAsBase64: absolute codeUri path', async () => {
-    await deploySupport.getFunCodeAsBase64('/a/b', process.cwd() + '/a/b/c/index.js');
-    assert.calledWith(lstat, process.cwd() + '/a/b/c/index.js');
-    assert.calledWith(zip.pack, process.cwd() + '/a/b/c/index.js', null);
-  });
-
-  it('test getFunCodeAsBase64: relative codeUri path', async () => {
-    await deploySupport.getFunCodeAsBase64('/a/b', './index.js');
-    assert.calledWith(lstat, './index.js');
-    assert.calledWith(zip.pack, './index.js', null);
-  });
-  
-});
-
-describe('test config invovationRole fot trigger', ()=> {
-  let restoreProcess;
-
-  beforeEach(async () => {
-    sandbox.stub(FC.prototype, 'getTrigger').resolves(undefined);
-    sandbox.stub(FC.prototype, 'createTrigger').resolves({});
-
-    Object.keys(ram).forEach(m => {
-      if (m === 'makeRole') {
-        sandbox.stub(ram, m).resolves({
-          'Role': {
-            'Arn': 'acs:ram::123:role/aliyunfcgeneratedrole-fc'
-          }
-        });
-      } else if (m === 'normalizeRoleOrPoliceName') {
-        sandbox.stub(ram, 'normalizeRoleOrPoliceName').returns('');
-      } else {
-        sandbox.stub(ram, m).resolves({});
-      }
-    });
-
-    deploySupport = await proxyquire('../../lib/deploy/deploy-support', {
-      '@alicloud/fc2': FC,
-      '../ram': ram
-    });
-
-    restoreProcess = setProcess({
-      ACCOUNT_ID: 'ACCOUNT_ID',
-      ACCESS_KEY_ID: 'ACCESS_KEY_ID',
-      ACCESS_KEY_SECRET: 'ACCESS_KEY_SECRET',
-      DEFAULT_REGION: 'cn-shanghai'
-    });
-  });
-
-  afterEach(() => {
-    sandbox.restore();
-    restoreProcess();
-  });
-
-  it('invocationRole  ots trigger ', async ()=> {
-    await deploySupport.makeTrigger(
-      {
-        serviceName: 'serviceName', 
-        functionName: 'functionName', 
-        triggerName: 'triggerName', 
-        triggerType: 'TableStore',
-        triggerProperties: {
-          'InstanceName': 'fc-test-inst',
-          'TableName': 'fc_test_tbl',
-          'InvocationRole': 'acs:ram::987:role/invocation-role-test'
-        }
-      });    
-
-    assert.notCalled(ram.makeRole);
-    assert.notCalled(ram.makePolicy);
-    assert.notCalled(ram.makeAndAttachPolicy);
-    assert.notCalled(ram.attachPolicyToRole);  
-
-    assert.calledWith(
-      FC.prototype.createTrigger,
-      'serviceName',
-      'functionName',
-      {
-        'triggerName': 'triggerName',
-        'triggerType': 'tablestore',
-        'triggerConfig': {},
-        'invocationRole': 'acs:ram::987:role/invocation-role-test',
-        'sourceArn': 'acs:ots:cn-shanghai:ACCOUNT_ID:instance/fc-test-inst/table/fc_test_tbl'
-
-      });
-  });
-});
